pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spotify-insights"
version = "2.0.0"
description = "Analyze your Spotify listening history with rich terminal output"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Zachariah Michael Lagden" }]
keywords = ["spotify", "music", "analytics", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "rich>=13.0",
    "click>=8.0",
    "spotipy>=2.23",
    "python-dotenv>=1.0",
    "textual>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "pyinstaller>=6.0",
]

[project.scripts]
spotify-insights = "spotify_insights.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["spotify_insights"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
